requirements.txt
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
Django==2.0.1
django-widget-tweaks==1.4.2
idna==2.6
Mastodon.py==1.2.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
